[CI][Nightly] tell a stuck benchmark file apart from an expensive one
The nightly benchmark job was cancelled at its 150-minute limit having run
34 of 59 files. Because GitHub cancelled it, bench_results.xml was never
written and publish-bench-data skipped: 149 minutes of exclusive GPU time
produced no data at all.

run_benchmarks.py had one knob, --timeout-per-file, doing two jobs. It
bounds a native failure, and it also kills a file for being expensive.
Three files reached the 900s cap while running normally -- bench_gqa.py
completes 34 tests in 915s cold-cache with its slowest single test at 76s.
The cap was self-perpetuating too: a file killed mid-run never writes its
kernels to the persistent TILELANG_CACHE_DIR, so it starts cold the next
night and reaches the cap again.

Two limits with separate jobs replace the one. --stall-timeout measures
time since the child last started a test, via a pytest plugin that rewrites
a heartbeat file the parent stats in the poll loop it already runs; a hang
is still caught as fast as before, and the kill message names the test.
--total-budget stops launching files, writes the merged report, and records
what it never reached as skipped, instead of letting CI cancel the job and
discard everything.

Also fixes three defects the same run exposed.

The child wrote its exit code to the status pipe before flushing stdout.
Redirected to a file, stdout is block-buffered, so the parent printed the
log while pytest's FAILURES section was still in the child's buffer:
nightly reported 12 failures with no reason attached to any of them.

Those 12: bench_binary_elementwise.py and bench_fused_moe_experts.py passed
dtype= to constructors that bind dtype per call, and
_ArgreduceOpBase._prepare_input skipped the shared input validation on its
strided fast path, so the op returned a correct result and then failed
eval_roofline() for an unbound dtype.

benchmarks/tests/test_run_benchmarks.py picks up the flag rename and three
cases the split introduces: a slow file outliving the stall timeout, a
failure report surviving the child's stdout buffer, and a spent budget
naming what it never reached.

+++ benchmarks/ops/bench_binary_elementwise.py
@@ -191,7 +191,7 @@ def test_comparison_bench(
     bm = BinaryBenchmark(test)
     inputs = test.gen_inputs()
 
-    op = _CMP_OPS[op_name](a_shape=shape, b_shape=shape, dtype=dtype)
+    op = _CMP_OPS[op_name](a_shape=shape, b_shape=shape)
     result = bm.profile(op, *inputs)
     BenchmarkReport.record(op, locals(), result, tag="tileops")
